reworked commands similar to how kubectl works
